test: adjust unit tests for FileSideBar
